add astype in coordinates matching

+++ deepsensor/data/loader.py
@@ -721,8 +721,8 @@ class TaskLoader:
             Y_c = df.values.T
         elif isinstance(sampling_strat, np.ndarray):
             X_c = sampling_strat.astype(self.dtype)
-            x1match = np.in1d(df.index.get_level_values("x1"), X_c[0])
-            x2match = np.in1d(df.index.get_level_values("x2"), X_c[1])
+            x1match = np.in1d(df.index.get_level_values("x1").astype(self.dtype), X_c[0])
+            x2match = np.in1d(df.index.get_level_values("x2").astype(self.dtype), X_c[1])
             num_matches = np.sum(x1match & x2match)
 
             # Check that we got all the samples we asked for
